Refactor para tratar operaciones con BigDecimal en lugar de Integer

+++ src/main/java/com/profile/test4/controller/CalculadoraController.java
@@ -6,6 +6,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 /**
  * The Class CalculadoraController.
  */
@@ -19,13 +21,13 @@ public class CalculadoraController {
 
     @GetMapping(value = "/suma/{firstNumber}/{secondNumber}", produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseStatus(HttpStatus.OK)
-    public Integer Suma(@PathVariable(required = true) Integer firstNumber, @PathVariable(required = true) Integer secondNumber){
+    public BigDecimal Suma(@PathVariable(required = true) BigDecimal firstNumber, @PathVariable(required = true) BigDecimal secondNumber){
         return calculadoraService.suma(firstNumber,secondNumber);
     }
 
     @GetMapping(value = "/resta/{firstNumber}/{secondNumber}", produces = MediaType.APPLICATION_JSON_VALUE)
     @ResponseStatus(HttpStatus.OK)
-    public Integer Resta(@PathVariable(required = true) Integer firstNumber, @PathVariable(required = true) Integer secondNumber){
+    public BigDecimal Resta(@PathVariable(required = true) BigDecimal firstNumber, @PathVariable(required = true) BigDecimal secondNumber){
         return calculadoraService.resta(firstNumber,secondNumber);
     }
 }
